Report on E-Access
The main aim of this project is to provide a simple messenger like platform using GUI. We
will use a simple programming language like Java and develop the modules in this project. 
Java contains certain swings which contain some methods that would be easier to design a 
project with ease of code. GUI is a user interface that allows us to interact with devices 
through graphical icons and visual indicators such as secondary notations, instead of text 
based user interface or text navigation. The actions in a Graphical User Interface are usually 
performed by direct manipulation of the graphical elements. GUIs are used in many mobile 
devices like MP3 players, portable players, gaming devices, phones and small household, 
office or industrial controls. It is not applied to lower-display resolution of interfaces like 
video games not including flat screens, like volumetric displays because it is restricted to 
two-dimensional display screen to describe generic information, in the tradition of the 
computer science research at the Xerox Palo Alto Research Center. By the 1980s, mobile
phones and game systems are also employed application specific touch screen GUIs. New 
automobiles use GUIs in multimedia centers, or navigation multimedia center combinations