report on parking slots
The main aim of the project is to book the slots of parking a vehicle for 
required time where fair charges are pressed for allotment as to avoid the 
mess in parking lot. It manages the details of duration, vehicles, parking 
slots and parking fares. It manages all the information about parking 
duration, slots and parking fare. The project is totally built at administrative 
end and thus only the administrator is guaranteed the access. The purpose 
of the project is to build an application program to reduce the manual work 
for managing the things. There will be a separate category for faculty and 
students to park the vehicle