Report on telecom billing
he project is mainly aimed to present the requirement of the Telephone Billing. In this 
project the telephone bills are automatically calculated. In this project almost every work is 
done that is related to telephone billing system. Telecom billing provides customer with many 
services like adding records, listing records, deleting records, record modification, viewing 
customer records & all works related to bills. Customer will also be provided with payment 
option. Some other options like Meter readings to bill calculation and bill generation are 
provided. “Telephone Billing” is been developed due to the increase in requirement of the 
work and to develop new work culture. Thus, to overcome such issues and problems a new 
software has been developed to minimize manual work and improve work efficiency by saving 
time and to provide flexibility and user-friendly environment to the customer. As the system 
that was proposed previously was very difficult and was manual with lot of errors this project 
was developed.